+++ src/slskd/Core/Data/Migrator.cs
@@ -58,6 +58,7 @@ public class Migrator
             { nameof(Z07152026_TransferUpdatedAtMigration), new Z07152026_TransferUpdatedAtMigration(connectionStrings: Databases) },
             { nameof(Z07152026_TransferHistoryIndexesMigration), new Z07152026_TransferHistoryIndexesMigration(connectionStrings: Databases) },
             { nameof(Z07162026_AutoRetryIndexMigration), new Z07162026_AutoRetryIndexMigration(connectionStrings: Databases) },
+            { nameof(Z07162026_DownloadRequestSummaryIndexMigration), new Z07162026_DownloadRequestSummaryIndexMigration(connectionStrings: Databases) },
         };
     }